chore: Update user login functionality and dependencies
Refactor the User class in user.js to improve the login functionality by removing unnecessary console.log statements and updating the storage of user data using SecureStore. Also, update the dependencies in package.json to use the latest stable version of Expo (51.0.4). Additionally, update the imports in UserAreaStack.js to reflect the new file structure. Finally, update the base row URL and table ID in .env.example to match the required environment variables for the app's integration with the Baserow API.