chore: Remove unused files and update dependencies
Update package.json to include the latest stable versions of "@shopify/flash-list" and "expo-dev-client". Remove the unused "invoices.js" file from the services directory. Update the "AssasController.js" file to include a new method for fetching bills from the Asaas API. Update the "FarturasStack.js" file to fetch and display bills in the "Faturas" tab of the user area.